+++ app/Listeners/Product/UpdateCacheProducts.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Listeners\Product;
+
+use App\Events\Product\ProductCreated;
+use Illuminate\Support\Facades\Cache;
+
+class UpdateCacheProducts
+{
+    /**
+     * Create the event listener.
+     */
+    public function __construct()
+    {
+        //
+    }
+
+    /**
+     * Handle the event.
+     */
+    public function handle(ProductCreated $event): void
+    {
+        Cache::tags(['products'])->flush();
+    }
+}