+++ app/Services/ProductService.php
@@ -1,21 +1,17 @@
 <?php
 namespace App\Services;
 
-use Cloudinary\Api\Exception\ApiError;
-
-use Illuminate\Support\Facades\Log;
-use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Support\Facades\Auth;
-use Cloudinary\Cloudinary;
-use Illuminate\Support\Facades\Cache;
-use Illuminate\Support\Carbon;
-
-use App\Events\ProductCreated;
-
+use App\Events\Product\ProductCreated;
 use App\Models\Filter;
 use App\Models\Product;
 use App\Models\Review;
 use App\Models\User;
+use Cloudinary\Api\Exception\ApiError;
+use Cloudinary\Cloudinary;
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Facades\Log;
 
 class ProductService {
     public function upload(array $validated_data, User $user) : void {